Make the following changes:
,

Find this element:

  <Card className="w-full h-64 flex flex-col items-center justify-center p-4 hover:shadow-md transition-shadow">

with the text content:

    Verwalten Sie Ihre vorhandenen Tests.

Relevant files:

    - _c@https://6000-idx-studio-1744472552773.cluster-6frnii43o5blcu522sivebzpii.cloudworkstations.dev/_next/static/chunks/src_components_5612caa1._.js
    - TeacherDashboard@https://6000-idx-studio-1744472552773.cluster-6frnii43o5blcu522sivebzpii.cloudworkstations.dev/_next/static/chunks/src_components_5612caa1._.js
    - react-stack-bottom-frame@https://6000-idx-studio-1744472552773.cluster-6frnii43o5blcu522sivebzpii.cloudworkstations.dev/_next/static/chunks/node_modules_next_dist_compiled_2ce9398a._.js

And change it as follows:

Makes this and all other cards square shaped and put them in a responsive grid, please. Make sure it resembles a kind of ios springboard with hige icons in the center. If you can, give each tile a different background color that fits our overall color theming. Make sure the contrast is always good in dark and light mode